tests/wpt/web-platform-tests/css/css-values/urls/empty.html

<!doctype html>
<title>Empty URLs behaviour</title>
<link rel=help href=https://drafts.csswg.org/css-values/#url-empty>
<link rel=help href=https://github.com/w3c/csswg-drafts/issues/2211#issuecomment-365677844>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
#inline-unquoted {
    background-image: url();
}

#inline-quoted {
    background-image: url("");
}
</style>
<link rel=stylesheet href=support/empty-urls.css>
<div id="inline-unquoted"></div>
<div id="inline-quoted"></div>
<div id="external-unquoted"></div>
<div id="external-quoted"></div>
<script>
const ids = [
  "inline-unquoted",
  "inline-quoted",
  "external-unquoted",
  "external-quoted"
];

const inline_url = location.href;
const external_url = new URL(document.querySelector("link[rel=stylesheet]").href, location.href).href;

for (let id of ids) {
    test(function() {
        const el = document.getElementById(id);
        const expected = id.startsWith("inline-") ? inline_url : external_url;
        const style = window.getComputedStyle(el);
        assert_equals(style["background-image"], 'url("' + expected + '")');
    }, "empty URL: " + id);
}
</script>
